Improve cantContain check
It was checking if any of the cantContain strings matched the line instead of the other way around. This method uses CharSequences instead and Java 8 streams to check if any line contains any of the cantContain strings but is not as limited. For example, you can enter "town_" in the config and it will match "town_funtown".